requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
streamlit
ultralytics
